Move getCookieParser into separate file to fix unit tests
