+++ packages/next/src/server/api-utils/index.ts
@@ -19,25 +19,6 @@ export type __ApiPreviewProps = {
 }
 
 /**
- * Parse cookies from the `headers` of request
- * @param req request object
- */
-export function getCookieParser(headers: {
-  [key: string]: string | string[] | null | undefined
-}): () => NextApiRequestCookies {
-  return function parseCookie(): NextApiRequestCookies {
-    const { cookie } = headers
-
-    if (!cookie) {
-      return {}
-    }
-
-    const { parse: parseCookieFn } = require('next/dist/compiled/cookie')
-    return parseCookieFn(Array.isArray(cookie) ? cookie.join('; ') : cookie)
-  }
-}
-
-/**
  *
  * @param res response object
  * @param statusCode `HTTP` status code of response
